requirements.txt
openai
kafka-python
python-dotenv
